[TASK] Upgrade for TYPO3 v13
The MjmlBasedView is replaced by an extension of the FluidEmail class which
renders the configured Fluid/Mjml templates by executing the mjml CLI command
in the usual manner.

For the now mandatory plain text variant of an email, an external HTML-to-text
converter library is used. Its integration as interface/service shall enable
additional pre- and post-processing steps within own extensions.